+++ src/game.h
@@ -1,6 +1,8 @@
 #ifndef GAME_H
 #define GAME_H
 
+#include "debug.h"
+
 void game_loop();
 
 #endif